+++ Libraries/pid.h
@@ -69,7 +69,7 @@ typedef struct {
 
     uint8_t idx;
     int16_t err[HISTORY_DATA_SIZE];
-    int16_t integrator;
+    int32_t integrator;
 
     int16_t     low_lim;
     int16_t     high_lim;
@@ -85,7 +85,7 @@ typedef struct {
  * @param lim  absolute limit range
  * @return clipped value
  */
-static int16_t abs_limit(int16_t *data, int16_t lim);
+static int32_t abs_limit(int32_t *data, int32_t lim);
 
 /**
  * @brief clip floating point number into [-lim, lim]
@@ -127,7 +127,7 @@ static float position_pid_calc(pid_ctl_t *pid);
  * @param dt        delta t in millisecond
  */
 void pid_init(pid_ctl_t *pid, pid_mode_t mode, motor_t *motor, 
-        int16_t low_lim, int16_t high_lim, int16_t int_lim, int16_t int_rng, int16_t max_derr,
+        int16_t low_lim, int16_t high_lim, int32_t int_lim, int16_t int_rng, int16_t max_derr,
         float kp, float ki, float kd, float maxout, float dt, float deadband);
 
 /**
